Extracts card from Links and makes separate LinkCard component
